# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "prod - hubApp",
    "description": "Production url endpoints",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://archi-hub-backend.vercel.app",
      "variables": {}
    }
  ],
  "paths": {
...
    "/auth/preSignUp": {
      "post": {
        "tags": ["auth"],
        "summary": "presignUp",
        "operationId": "presignUp",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/presignUpRequest"
                  },
                  {
                    "example": {
                      "email": "micheal.badmus@gmail.com",
                      "password": "Mhyke"
                    }
                  }
                ]
              },
              "example": {
                "email": "micheal.badmus@gmail.com",
                "password": "Mhyke"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "presignUpRequest": {
        "title": "presignUpRequest",
        "required": ["email", "password"],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "email": "micheal.badmus@gmail.com",
          "password": "Mhyke"
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}